adjust osd_get_max_rows() to return type int instead of uint8_t so that we don't need to include <stdint.h>

+++ SilverLite/drv_osd.cpp
@@ -544,7 +544,7 @@ void osd_clear()
 }
 
 //------------------------------------------------------------------------------
-uint8_t osd_get_max_rows()
+int osd_get_max_rows()
 {
-	return maxScreenRows;
+	return (int)maxScreenRows;
 }
